fixed 16

+++ 16-binary_tree_is_perfect.c
@@ -33,13 +33,12 @@ int is_perfect(const binary_tree_t *tree, int height)
  *
  * Return: Height of the tree
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t  binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
-	int left_height = binary_tree_height(tree->left);
-	int right_height = binary_tree_height(tree->right);
+	size_t left_height = binary_tree_height(tree->left);
+	size_t right_height = binary_tree_height(tree->right);
 
 	return ((left_height > right_height ? left_height : right_height) + 1);
 }
-
